1.5 Se agilizo el codigo
Se optimizo el codigo
