+++ BACKEND/controllers/authController/logoutUser.js
@@ -1,0 +1,4 @@
+//Cerrar session(aun no funka xd)
+exports.logout = (req, res) => {
+    res.json({ message: "logout exitoso" });
+  };
